requirements.txt
django==2.2.10
python-dotenv==0.12.0
djangorestframework